requirements.txt
annotated-types
anyio
arxiv
asgiref
beautifulsoup4
bibtexparser
bs4
certifi
cffi
charset-normalizer
colorama
contextlib2
copyleaks
cryptography
defusedxml
distro
Django
django-cors-headers
djangorestframework
docopt
feedparser
h11
httpcore
httpx
idna
internetarchive
jsonpatch
jsonpointer
langchain-core
langchain-openai
langsmith
oauthlib
openai
orjson
packaging
pycparser
pydantic
pydantic_core
PyJWT
pyparsing
python-dateutil
python-social-auth
python3-openid
pytz
PyYAML
pyzotero
regex
requests
requests-oauthlib
resend
schema
sgmllib3k
six
sniffio
social-auth-app-django
social-auth-core
soupsieve
sqlparse
tenacity
tiktoken
tqdm
typing_extensions
tzdata
urllib3
